# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "EBYS - İş Akışı ve Onay Yönetim Sistemi API",
    "description" : "Uçların çoğu kimlik doğrulaması ister.\n\nKullanım: `/api/auth/login` ucundan e-posta ve şifreyle giriş yapın,\ndönen `accessToken` değerini sağ üstteki **Authorize** butonuna yapıştırın.\nBundan sonraki isteklere `Authorization: Bearer <token>` başlığı otomatik eklenir.\n",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/auth/refresh" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "refresh",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "RefreshTokenRequest" : {
        "type" : "object",
        "properties" : {
          "refreshToken" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "refreshToken" ]
      },
      "LoginResponse" : {
        "type" : "object",
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "mustChangePassword" : {
            "type" : "boolean"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Giriş sonrası dönen accessToken",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}